// src/app/api-client/model/adminAnalyticsDTO.ts
/**
 * TeachingBACKEND API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { GeographicStatsDTO } from './geographicStatsDTO';
import { QuizPerformanceDTO } from './quizPerformanceDTO';
import { UserGrowthDTO } from './userGrowthDTO';
import { LearnHubPerformanceDTO } from './learnHubPerformanceDTO';
import { MonthlyRevenueDTO } from './monthlyRevenueDTO';


export interface AdminAnalyticsDTO { 
    monthlyRevenue?: Array<MonthlyRevenueDTO> | null;
    userGrowth?: Array<UserGrowthDTO> | null;
    topPerformingLearnHubs?: Array<LearnHubPerformanceDTO> | null;
    mostChallengingQuizzes?: Array<QuizPerformanceDTO> | null;
    geographicDistribution?: Array<GeographicStatsDTO> | null;
}