// src/app/api-client/model/adminStatsDTO.ts
/**
 * TeachingBACKEND API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AdminStatsDTO { 
    activeLearnHubs?: number;
    totalRegisteredUsers?: number;
    completedQuizzes?: number;
    activeAdministrators?: number;
    totalSchools?: number;
    totalStudents?: number;
    totalFamilies?: number;
    pendingApprovals?: number;
    totalRevenue?: number;
    activeSubscriptions?: number;
}