// src/app/api-client/model/monthlyRevenueDTO.ts
/**
 * TeachingBACKEND API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface MonthlyRevenueDTO { 
    year?: number;
    month?: number;
    revenue?: number;
    newSubscriptions?: number;
    cancellations?: number;
}